feat: implement DUST GraphQL API with complete functionality
